feat(rust): Add bolero generators for gRPC types
TypeGenerators need to be in this crate due to the rust orphan
rule which states that only crates that implement the trait or
the type can implement a trait for a given type.  And, since
TypeGenerator is in the bolero crate, and the proto types are
here, we have to implement these here.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>