fix(ci): Use rust 1.86.0 specifically
Rust 1.86.0 does not have unbounded_shl/shr code.  Here there is a
workaround for 1.86.0 that will generate an error on latest stable
where the unbounded shifts are stable.  When we update everything
to 1.87.0 in other repos, this can be changed to the appropriate
version.

We will get an error when we make the update so we can remove
the workaround.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>