chore(rust): Add CI workflow to bump Rust dependencies
Dependabot manages Go dependencies for the repository, but it's not
currently set up to take care of Rust dependencies.

We could set up Dependabot to take care of Rust too, I think; here we
simply reuse the workflow we have for the dataplane repository, because
it's been working well for us.

This workflow does not contain all steps from the original one; in
particular, we don't run `cargo deny check` before or after the updates,
because cargo-deny has not been configured for this repository.

Link: https://github.com/githedgehog/dataplane/blob/2c28bbfcedf1/.github/workflows/bump.yml
Signed-off-by: Quentin Monnet <qmo@qmon.net>