feat: Add ability to use well know types in .proto
Config for NAT will use google.protobuf.Duration
but our setup for prost and protobuf did not allow
this.  This commit fixes that problem.

The normal way to do this is to use prost-types
but, prost-types do not derive serde::Serialize
and serde::Deserialize, so this commit instead
has prost generate the types directly.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>